refactor(techdocs): extract reader context
