+++ plugins/techdocs/src/reader/components/Reader/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './Reader';
+export * from './context';
